mlanguage: add a matching_expr_state relation and properties wrt split_state
The core idea is that for the linking/wrapping semantics, requiring
all states to be splittable according to split_state is too demanding.

In other words, given σ a state of the linking/wrapping semantics,
   ∃ pubσ privσ, split_state σ pubσ privσ
will only be true for a some specific class of σs.

In practice, this class of splittable σ informally corresponds to
"external σs" that are found at the boundary of a module, i.e when
receiving an incoming function call, or when producing an outgoing
value. "Internal σs" which appear at the middle of reduction inside a
linked/wrapped program are typically not splittable.

[matching_expr_state] captures this additional relation between states
and "shape" of the program. Then:

- [matching_expr_state e σ] is intended to hold of all pairs (e, σ) of
  an execution (it is preserved by reduction)
- if we can see that e corresponds to a program boundary (incoming
  function call or outgoing value), then [matching_expr_state e σ]
  is equivalent to saying that σ is splittable.