Make the [language] interface more abstract
Move head_step & low-level property on evaluation contexts "under" the
abstraction barrier of the language interface. The goal is to make
[language] closer to [mlanguage], ando only expose properties that we
really need for the rest of the setup, in order to make it easier to
make change to the individual languages' semantics.

We also get rid of [wp_bind_inv] (that we did not use), which allows
us to stop proving some tricky properties (that might not be very
robust wrt future language extensions).